Move DA Boundary Checks to CustodyContext
